addresses represented with 64 bits

+++ include/retdec/fileformat/types/tls_info/tls_info.h
@@ -19,11 +19,11 @@ namespace fileformat {
 class TlsInfo
 {
 	private:
-		std::vector<std::uint32_t> callBacks;  ///< addresses of callback functions
-		std::uint32_t rawDataStartAddr;        ///< start address of raw data
-		std::uint32_t rawDataEndAddr;          ///< end address of raw data
-		std::uint32_t indexAddr;               ///< address of index
-		std::uint32_t callBacksAddr;           ///< address of array of callbacks
+		std::vector<std::uint64_t> callBacks;  ///< addresses of callback functions
+		std::uint64_t rawDataStartAddr;        ///< start address of raw data
+		std::uint64_t rawDataEndAddr;          ///< end address of raw data
+		std::uint64_t indexAddr;               ///< address of index
+		std::uint64_t callBacksAddr;           ///< address of array of callbacks
 		std::uint32_t zeroFillSize;            ///< size of zero fill
 		std::uint32_t characteristics;         ///< characteristics
 		bool rawDataStartAddrValid;            ///< member validity flag
@@ -38,28 +38,28 @@ class TlsInfo
 
 		/// @name Getters
 		/// @{
-		bool getRawDataStartAddr(std::uint32_t &res) const;
-		bool getRawDataEndAddr(std::uint32_t &res) const;
-		bool getIndexAddr(std::uint32_t &res) const;
-		bool getCallBacksAddr(std::uint32_t &res) const;
+		bool getRawDataStartAddr(std::uint64_t &res) const;
+		bool getRawDataEndAddr(std::uint64_t &res) const;
+		bool getIndexAddr(std::uint64_t &res) const;
+		bool getCallBacksAddr(std::uint64_t &res) const;
 		bool getZeroFillSize(std::uint32_t &res) const;
 		bool getCharacteristics(std::uint32_t &res) const;
-		const std::vector<std::uint32_t> &getCallBacks() const;
+		const std::vector<std::uint64_t> &getCallBacks() const;
 		/// @}
 
 		/// @name Setters
 		/// @{
-		void setRawDataStartAddr(std::uint32_t sAddr);
-		void setRawDataEndAddr(std::uint32_t eAddr);
-		void setIndexAddr(std::uint32_t iAddr);
-		void setCallBacksAddr(std::uint32_t cbAddr);
+		void setRawDataStartAddr(std::uint64_t sAddr);
+		void setRawDataEndAddr(std::uint64_t eAddr);
+		void setIndexAddr(std::uint64_t iAddr);
+		void setCallBacksAddr(std::uint64_t cbAddr);
 		void setZeroFillSize(std::uint32_t zFill);
 		void setCharacteristics(std::uint32_t chars);
 		/// @}
 
 		/// @name Other methods
 		/// @{
-		void addCallBack(std::uint32_t cb);
+		void addCallBack(std::uint64_t cb);
 		/// @}
 };
